move TCP DNS Server timeout to dnsutils folder
