+++ pkg/networkservice/chains/nsmgr/dns_test.go
@@ -21,12 +21,10 @@ package nsmgr_test
 
 import (
 	"context"
-	"fmt"
 	"net"
 	"testing"
 	"time"
 
-	"github.com/miekg/dns"
 	"github.com/networkservicemesh/api/pkg/api/networkservice"
 	"github.com/networkservicemesh/api/pkg/api/networkservice/mechanisms/cls"
 	kernelmech "github.com/networkservicemesh/api/pkg/api/networkservice/mechanisms/kernel"
@@ -143,141 +141,5 @@ func Test_DNSUsecase(t *testing.T) {
 	require.NoError(t, err)
 
 	_, err = nse.Unregister(ctx, nseReg)
-	require.NoError(t, err)
-}
-
-type proxyDNSServer struct {
-	ListenOn  string
-	TCPServer *dns.Server
-	UDPServer *dns.Server
-}
-
-func (p *proxyDNSServer) listenAndServe(t *testing.T, tcpHandler, udpHandler dns.Handler) {
-	p.TCPServer = &dns.Server{Addr: p.ListenOn, Net: "tcp", Handler: tcpHandler}
-	p.UDPServer = &dns.Server{Addr: p.ListenOn, Net: "udp", Handler: udpHandler}
-
-	go func() {
-		err := p.TCPServer.ListenAndServe()
-		require.NoError(t, err)
-	}()
-
-	go func() {
-		err := p.UDPServer.ListenAndServe()
-		require.NoError(t, err)
-	}()
-}
-
-func (p *proxyDNSServer) shutdown() error {
-	tcpErr := p.TCPServer.Shutdown()
-	udpErr := p.UDPServer.Shutdown()
-
-	if tcpErr != nil {
-		return tcpErr
-	}
-
-	if udpErr != nil {
-		return udpErr
-	}
-
-	return nil
-}
-
-type tcpHandler struct {
-}
-
-func (h *tcpHandler) ServeDNS(rw dns.ResponseWriter, m *dns.Msg) {
-	time.Sleep(time.Second * 5)
-	dns.HandleFailed(rw, m)
-}
-
-type udpHandler struct {
-}
-
-func (h *udpHandler) ServeDNS(rw dns.ResponseWriter, m *dns.Msg) {
-	if m.Question[0].Name == "my.domain." {
-		dns.HandleFailed(rw, m)
-		return
-	}
-
-	name := dns.Name(m.Question[0].Name).String()
-	rr := new(dns.A)
-	rr.Hdr = dns.RR_Header{Name: name, Rrtype: dns.TypeA, Class: dns.ClassINET, Ttl: 3600}
-	rr.A = net.ParseIP("1.1.1.1")
-
-	resp := new(dns.Msg)
-	resp.SetReply(m)
-	resp.Authoritative = true
-	resp.Answer = append(resp.Answer, rr)
-
-	if err := rw.WriteMsg(resp); err != nil {
-		dns.HandleFailed(rw, m)
-	}
-}
-
-func getFreePort() (int, error) {
-	addr, err := net.ResolveTCPAddr("tcp", "localhost:0")
-	if err != nil {
-		return 0, err
-	}
-
-	l, err := net.ListenTCP("tcp", addr)
-	if err != nil {
-		return 0, err
-	}
-	err = l.Close()
-	if err != nil {
-		return 0, err
-	}
-
-	return l.Addr().(*net.TCPAddr).Port, nil
-}
-
-func Test_TCPDNSServerTimeout(t *testing.T) {
-	t.Cleanup(func() { goleak.VerifyNone(t) })
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
-	defer cancel()
-
-	proxyPort, err := getFreePort()
-	require.NoError(t, err)
-	proxyAddr := fmt.Sprintf("127.0.0.1:%d", proxyPort)
-
-	proxy := &proxyDNSServer{ListenOn: proxyAddr}
-	proxy.listenAndServe(t, &tcpHandler{}, &udpHandler{})
-
-	clientPort, err := getFreePort()
-	require.NoError(t, err)
-	clientAddr := fmt.Sprintf("127.0.0.1:%d", clientPort)
-
-	dnsConfigsMap := new(dnsconfig.Map)
-	dnsConfigsMap.Store("1", []*networkservice.DNSConfig{
-		{
-			DnsServerIps:  []string{proxyAddr},
-			SearchDomains: []string{"com"},
-		},
-	})
-
-	clientDNSHandler := next.NewDNSHandler(
-		dnsconfigs.NewDNSHandler(dnsConfigsMap),
-		searches.NewDNSHandler(),
-		noloop.NewDNSHandler(),
-		fanout.NewDNSHandler(),
-	)
-	dnsutils.ListenAndServe(ctx, clientDNSHandler, clientAddr)
-
-	resolver := net.Resolver{
-		PreferGo: true,
-		Dial: func(ctx context.Context, network, address string) (net.Conn, error) {
-			var dialer net.Dialer
-			return dialer.DialContext(ctx, network, clientAddr)
-		},
-	}
-
-	resolveCtx, resolveCancel := context.WithTimeout(context.Background(), time.Second*5)
-	defer resolveCancel()
-
-	requireIPv4Lookup(resolveCtx, t, &resolver, "my.domain", "1.1.1.1")
-
-	err = proxy.shutdown()
 	require.NoError(t, err)
 }
